requirements.txt
amqp==2.5.2
ansible==2.7.4
asgiref==3.2.5
bcrypt==3.1.7
billiard==3.6.3.0
celery==4.4.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
colorlog==5.0.1
cryptography
Django==2.1.8
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
elasticsearch==7.12.0
flower==0.9.7
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
itypes==1.1.0
Jinja2==2.11.1
kombu==4.6.8
MarkupSafe==1.1.1
openapi-codec==1.3.2
paramiko==2.7.1
pika==1.1.0
pycparser==2.20
PyNaCl==1.3.0
pytz==2019.3
PyYAML==5.3.1
PyJWT==1.7.1
redis==3.4.1
requests==2.23.0
salt==3000.3 
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
supervisor==4.2.0
timelib==0.2.5
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
websocket-client==0.57.0
zipp==3.1.0
nacos-sdk-python==0.1.5
kubernetes==12.0.0 
redfish